+++ src/utils.ts
@@ -1,7 +1,9 @@
 import { treeGraph } from './tree-graph.js';
 import exitWithError from './exit-with-error.js';
+import countSubcommands from './utils/count-commands.js';
 
 export {
     exitWithError,
     treeGraph,
+    countSubcommands,
 };