+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Updated tree graph command error message to include the command of the subcommand
 - Updated usage in [README](README.md)
+- Added a function to count subcommands
 
 ## v2.1.0 2024-07-18